k32w0: cbx: add PTA target generation for RCPs
For the OT RCP targets with flow control and with UART DMA,
as well as for the OT/BLE single UART hybrid examples, build
the corresponding PTA target as well. This is done in a separate
directory so it can be easily traced.

Signed-off-by: Alex Porosanu <alexandru.porosanu@nxp.com>
(cherry picked from commit 9d8858cf17c135ca927ffe2a312fff1c71184804)