k32w1: hybrid: rework ot-rcp-zb-ncp example
This is a rework of 70fd18a3 that updates the hybrid
OT RCP/ZB NCP hybrid example so that it builds on top of the
latest SDK tree. Next, once the SDK generator is updated to contain
the related ZB NCP files, this target will be enabled for default build.

Signed-off-by: Alex Porosanu <alexandru.porosanu@nxp.com>