k32w1: examples: hybrid: ensure proper line-endings
Due to the fact that the K32W1 SDK ZIP package is built under Windows,
the .sh scripts used for the hybrid examples have Windows line endings
and thus fail to work on Linux machines. The simplest solution is to
force a conversion to Unix line endings whenever the hybrid examples
are built.

Signed-off-by: Alex Porosanu <alexandru.porosanu@nxp.com>