k32w1: examples: hybrid: ot_cli_ftd_zc: fix console component inclusion
Part of the new SDK K32W1 SDK release, we've split the UART functionality
from the console functionality (albeit they are identical in some cases).
As such, the example(s) need to be updated so they pull in the new `console.c`
file from the SDK sources.
This patch also adds the `app_console.c` file that's needed by the application(s)
to work with the newly added console abstraction.

While here, also update the call to UART_vInit() that now accepts a parameter.

Also, remove the compilation warnings due to the missing prototypes for
APP_taskAtSerial() &  APP_SerialCmdTask() functions.

Signed-off-by: Alex Porosanu <alexandru.porosanu@nxp.com>