k32w0: script: sync with the sign_images script changes
sign_images.sh script was modified to enabled support for additional arguments.
More specifically, it was introduced by the simple hash verification request and
it created a regression on the ot-cli apps, caused by the args misalignment. The
resulting binaries were not signed.

Introduced support for using an environment variable (NXP_SIGN_ADDITIONAL_ARGS) to
specify additional arguments. If not set, the command will pass an empty arg to the
sign command, which will not affect the output.

Signed-off-by: marius-alex-tache <marius.tache@nxp.com>