k32w0: fix OPENTHREAD_CONFIG_TMF_ADDRESS_CACHE_ENTRIES
As part of the upstream stack commit 52e157b17b4fcc7682c0f6ee1618ebb38674720b,
OPENTHREAD_CONFIG_TMF_ADDRESS_CACHE_ENTRIES is set to 256 is CONFIG_BORDER_ROUTER
is also set. For K32W061 this is set to 1, but we do not have the neccessary
resources (i.e. SRAM) to store these many entries, especially in the case of
the hybrid OT/ZB example. Since we don't want to change the options for
all the K32W061 examples, the simplest solution is to override the
setting for OPENTHREAD_CONFIG_TMF_ADDRESS_CACHE_ENTRIES only for
the OT FTD/ZB ZED example, and set it as before to 32.

Signed-off-by: Alex Porosanu <alexandru.porosanu@nxp.com>