RW612: Add workaround for A2 boards with eth workarounds
A2 rw612 boards that have eth workaround done in HW will fail at
initialization due to the fact that the external oscillator is not
available anymore.
In this case we need to bypass the code that inits the external
oscillator, and we do that by the specified define.
By default this setting is not define since it depends on the type of
board that the user is using and therefore should be initialized on a
need to use basis.