[otbr] Fixed memory leak
Fixed memory leak in LWIP caused by RAW socket usage.
When the ICMP RAW socket is created for sending OT stack generated RAs,
it's initialized with a generic raw receive function that puts all received IMCP packets
in a queue. This queue is never read and freed causing a memory leak.
The solution is to stop using the RAW socket since there is no API to overwrite
that function. Instead, we use the same RAW API that is used for receiving
IMCP packets for the OT stack. In this case, the receive function is already
initialized by us and we will not store any packet.

Signed-off-by: Marius Preda <marius.preda@nxp.com>