[common][border-router] Fix deadlock between LWIP and OT tasks
This commit fixes a deadlock between the LWIP and OT tasks that happens
when an ND packet generated by OT needs to be sent on the external interface.
This action was executed in the context of the OT task locking the LWIP task.
On some rare occasions, a packet received on the external interface would
lock the LWIP task before the OT task can take the lock. After the packet
was received on the external interface it would also be forwarded to the OT IP
layer causing the LWIP task to lock waiting for the OT that was already waiting
for the LWIP task. This causes a deadlock. The fix consists of sending the OT
generated ND packet on the context of the LWIP task using a message mechanism.
This removes the need to lock the LWIP task when sending this ND packet.

Signed-off-by: Marius Preda <marius.preda@nxp.com>