[border-router][common] Updated Border Router implementation for newer OpenThread stack.
- Improve otPlatUdpConnect function.

Improvements were done to accomodate the new socket.Connect call from
Openthread secure_transport.cpp.
In commissioning scenario, dataset could not be retrieved due to a
missmatch at lwip udp_input, as socket was bound to unspecified netif,
and destination unreachable was sent as a response when dtls handshake
was performed.

Also, otTaskCb is now not freeing the message, as new mdns
implementation takes ownership of an incoming message and does all the
handling, including freeing memory.

- Add InfraIfLinkState API

Added a new API in infra_if.c that allows informing the Routing Manager of the OT stack when
the link state changes. This API can then be called from the App interface manager to inform OT whenever
the link is up or down.

- Create utilitary file for platform modules.

This file was created to accomodate functions like setting
services/hosts names and setting alternative ones in case of conflicts.
It can be used to include future functionalities that can be shared
between modules.

- Updating Border Agent platform module to support new features.

> Provide host name at initialization which will be used to populate the mdns service structure.
> Added support for Ephemeral Key functionality.
> Use utilitary function to set base/alternative service name.
> Fix compilation warning reported in 'GetStateBitmap()'

- Refactorization and updates on br_rtos_manager module.

>  Created 'BrInitPlatform()'.
>  Use new routing manager API for processing DHCP6 prefix

- Added mdns_socket platform implementation required by mDNS module.

- Add config flags to reference new OT features.

> Enabled native OT mDNS module.
> Enabled advertising proxy.

Signed-off-by: Cristib05 <cristian.bulacu@nxp.com>