k32w0: examples: hybrid: add OT RCP/BLE with UART DMA
This patch replaces the standard USART flow control mechanism
that's not exposed over K32W0 USB connector with a DMA based one.
This is already used by the RCP UART DMA example.

Ideally this would allow the K32W0 dongle to be used as a combo
device that allow f.i. a IMX type of device be a MATTER comissioner
(over BLE) and border router (over 15.4) without any additional
hardware.

Signed-off-by: Alex Porosanu <alexandru.porosanu@nxp.com>