[common] Align Zephyr NVS OT settings key names to Matter key names
In the implementation of the settings storage layer using the Zephyr
NVS file system, the settings names format was "OT/xyz" (capitals)
while the NVS keys in Matter applications use the format "mt_i/xyz"
or "mt_s/xyz" (lowercase). This update changes the NVS settings key
names format in OpenThread to "ot/xyz" (lowercase) so that it aligns
with the NVS key format of the Matter applications.

Signed-off-by: Marian Chereji <marian.chereji@nxp.com>