[border-router] Fix DAD issue on BR Ethernet App.
Removed border router application hardcoded configMAC_ADDR define.
For RT1060 and RT1170, MAC address is computed using SILICONID_ConvertToMacAddr() function.
For RW612, as this functionality is not available, appGenerateMacAddr() was defined,
trying to replicate as much as possible from SILICONID_ConvertToMacAddr(), but it is using a random value generated with otPlatEntropyGet().

Signed-off-by: Cristib05 <cristian.bulacu@nxp.com>