[common] Fix failure to remove OT settings during factory reset
When using the NVM storage solution, occasionally, the OT settings
fail to be reinitialized during a device factory reset. This occurs
because the NvSyncSave() is being used while the
OT_SAVE_NVM_DATA_ON_IDLE feature is enabled and it occasionally
conflicts with NVM activities started by the idle task.

A temporary workaround was introduced until the NvShutdown() API
can be fixed to correctly execute all remaining NVM tasks in the
queue.

Signed-off-by: Marian Chereji <marian.chereji@nxp.com>
Reviewed-by: Doru-Cristian Gucea <doru-cristian.gucea@nxp.com>
Reviewed-by: Philippe Guasch <philippe.guasch@nxp.com>
Reviewed-by: George Stefan <george.stefan@nxp.com>
Reviewed-by: Marius Tache <marius.tache@nxp.com>
Reviewed-by: Andrei Menzopol <andrei.menzopol@nxp.com>
Verified-by: Forrest Sun <forrest.sun@nxp.com>