[border-router][common] Replace generic tasklet with a new implementation
This commit replaces the generic tasklet feature from Open Thread because it is not accepted upstream. Instead, a new function, otPlatUdpProcess, was created that handles the processing of platform UDP sockets in the context of the OT stack. A list is used to add received messages from the LWIP task and the processing of messages happens on the OT task.