[border-router][common] Fix udp_plat in release mode
The new method that handles the processing of platform UDP sockets in the context of the OT stack, otPlatUdpProcess, was locking on the mutex
during boot when building in release mode. The issue was caused by improper initialization of the mutex in the init function since it was wrapped
inside an assert statement that is not included in release mode.
The fix consists of moving out the assert the initialization of the mutex.
Also, a simplification was made to the otPlatUdpProcess function.

Signed-off-by: Cristian Bulacu <cristian.bulacu@nxp.com>