[common][nvs] Fix setting delete operation
This commit fixes the removal of a settings key, either one index or all the values for that key.
Previous implementation did not take into account correctly the settings index provided by OpenThread, resulting in the generation of incorrect NVS key string. This led to the delete function not working as the string used for deleting never matched the string used when the key was saved.

Signed-off-by: Marius Preda <marius.preda@nxp.com>