[k32w1][mcx7x] Fix CSL IE in TX packets
This commit fixes broken CSL functionality.
Previously aFrame->mInfo.mTxInfo.mCslPresent was used to determine if the radio code should update                    the IE header. This field is no longer set by the OT stack. Until the issue is fixed in OT stack check if CSL period is > 0 and always update CSL IE in that case.

Signed-off-by: Marius Preda <marius.preda@nxp.com>