[common][border-router] Enable multicast loop for discovery proxy
This commit enables multicast loop option in LWIP allowing multicast
queries generated by the discovery proxy mechanism to be looped back
in the BR MDNS responder and checked against BR's own services.
This fixes the case where a service/host that a thread node wants
to discover is located on the BR itself.
Without this feature the query will only be sent on the LAN and
the response will never be found.

Signed-off-by: Marius Preda <marius.preda@nxp.com>