[border-router][trel_plat] Fix issues related to initialization and peer address selection
This commit aims to fix an initialization issue, trel socket was not
properly bind to netif because call to this function was done before udp
plat was actually initialized and it's netif structures were populated.
Also, trel peer address choice was fixed, a higher scope address is
selected as described in spec.

Signed-off-by: Cristian Bulacu <cristian.bulacu@nxp.com>