[border-router] Fix memory corruption
This commit fixes a memory corruption when the number of IPv6 addresses on the external interface was higher than MAX_HOST_IPV6_ADDRESSES. The UpdateIp6AddressList function was ported form Matter and was adapted. In the original Matter code, there was a mistake in the condition that checked if the address array was full. This mistake became less obvious when the code was moved in ot-nxp.

Increased the number of IPv6 addresses registered by MDNS.

Signed-off-by: Marius Preda <marius.preda@nxp.com>