[rt1060][rt1170][rw612][radio] Updated fix for endianess in otPlatRadioClearSrcMatchExtEntry
This commit updates the endianness fix in otPlatRadioClearSrcMatchExtEntry
to allow using the current RCP firmware for IW/RW. The original fix did
not work without a change on both sides, host and rcp.
Since K32W0 can be used as both standalone and rcp, the fix takes this
into account when building a specific target.

Signed-off-by: Marius Preda <marius.preda@nxp.com>